Add the Proto_manager class.
The Proto_manager should be only one instance (a singleton) that "knows"
how to deal with protobufs. It will be responsible for converting them
back and forth from binary and text representations and it will keep track
of where the protobuf definitions for different columns are stored. For
example, right now they are stored in the column comment, but in the
future we want them stored in a separate file.

Change-Id: I63f760381ad767b8f02b4d64b5c5c11835aa87cd
Reviewed-on: https://speckle-internal-review.googlesource.com/1142
Reviewed-by: Razvan Musaloiu-E. <razvanm@google.com>
Tested-by: Razvan Musaloiu-E. <razvanm@google.com>